Move time series space view to its own crate
