+++ crates/re_space_view_time_series/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+authors.workspace = true
+description = "A Space View that shows plots over Rerun timelines."
+edition.workspace = true
+homepage.workspace = true
+license.workspace = true
+name = "re_space_view_time_series"
+publish = true
+readme = "README.md"
+repository.workspace = true
+rust-version.workspace = true
+version.workspace = true
+include = ["../../LICENSE-APACHE", "../../LICENSE-MIT", "**/*.rs", "Cargo.toml"]
+
+[package.metadata.docs.rs]
+all-features = true
+
+[dependencies]
+re_arrow_store.workspace = true
+re_components.workspace = true
+re_log_types.workspace = true
+re_log.workspace = true
+re_renderer.workspace = true
+re_space_view.workspace = true
+re_tracing.workspace = true
+re_ui.workspace = true
+re_viewer_context.workspace = true
+re_query.workspace = true
+re_format.workspace = true
+
+egui.workspace = true
+vec1.workspace = true
