+++ crates/re_space_view_time_series/src/lib.rs
@@ -1,0 +1,8 @@
+//! Rerun time series Space View
+//!
+//! A Space View that shows plots over Rerun timelines.
+
+mod scene_part;
+mod space_view_class;
+
+pub use space_view_class::TimeSeriesSpaceView;
