+++ crates/re_viewer/Cargo.toml
@@ -55,6 +55,7 @@ re_space_view_bar_chart.workspace = true
 re_space_view_spatial.workspace = true
 re_space_view_text_box.workspace = true
 re_space_view_text.workspace = true
+re_space_view_time_series.workspace = true
 re_time_panel.workspace = true
 re_tracing.workspace = true
 re_ui = { workspace = true, features = ["eframe"] }
