+++ crates/re_viewer/src/app.rs
@@ -940,10 +940,11 @@ impl eframe::App for App {
 fn populate_space_view_class_registry_with_builtin(
     space_view_class_registry: &mut SpaceViewClassRegistry,
 ) -> Result<(), SpaceViewClassRegistryError> {
-    space_view_class_registry.add::<re_space_view_text::TextSpaceView>()?;
-    space_view_class_registry.add::<re_space_view_text_box::TextBoxSpaceView>()?;
-    space_view_class_registry.add::<re_space_view_spatial::SpatialSpaceView>()?;
     space_view_class_registry.add::<re_space_view_bar_chart::BarChartSpaceView>()?;
+    space_view_class_registry.add::<re_space_view_spatial::SpatialSpaceView>()?;
+    space_view_class_registry.add::<re_space_view_text_box::TextBoxSpaceView>()?;
+    space_view_class_registry.add::<re_space_view_text::TextSpaceView>()?;
+    space_view_class_registry.add::<re_space_view_time_series::TimeSeriesSpaceView>()?;
     Ok(())
 }
