--- crates/re_viewport/src/view_time_series/mod.rs
@@ -1,5 +1,0 @@
-mod scene;
-pub(crate) use self::scene::SceneTimeSeries;
-
-mod ui;
-pub(crate) use self::ui::{help_text, view_time_series, ViewTimeSeriesState};
