+++ scripts/publish_crates.sh
@@ -122,9 +122,10 @@ cargo publish $FLAGS -p re_time_panel
 cargo publish $FLAGS -p re_space_view
 cargo publish $FLAGS -p re_viewport
 cargo publish $FLAGS -p re_space_view_bar_chart
+cargo publish $FLAGS -p re_space_view_spatial
 cargo publish $FLAGS -p re_space_view_text
 cargo publish $FLAGS -p re_space_view_text_box
-cargo publish $FLAGS -p re_space_view_spatial
+cargo publish $FLAGS -p re_space_view_time_series
 cargo publish $FLAGS -p re_viewer
 cargo publish $FLAGS -p re_sdk
 cargo publish $FLAGS -p rerun
